% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesistest.R
\name{ttest_chi}
\alias{ttest_chi}
\title{Provides the t-test and chi square's p value and statistic for binary targets and provides it in a dataframe for a set of columns or for a whole dataframe.In case of ANOVA provides all possible tests's summary}
\usage{
ttest_chi(data, type_of_test, User_BinaryTarget = NULL,
  User_Variable = NULL, filename = NULL)
}
\arguments{
\item{For}{Hyp_test()----->data,type_of_test,User_BinaryTarget(optional),User_Variable(optional)}
}
\description{
1.Provides the t-test and chi square's p value and statistic and provides it in a dataframe for a set of columns or for a whole dataframe.
   2.In case of ANOVA() provides p values in form of a dataframe
   Assumption: No individual columns are provided for function Hyp_test().In such case a normal one on one t-test or chi would be better.
}
\examples{
ttest_chi(df,"chi",c(1,2),c(3,4)),ttest_chi(df,"ttest",filename="Apple")
}
